+++ docs/source/whatsnew.rst
@@ -29,6 +29,8 @@ Legend
 
 Latest
 ======
+- |Add| code to download :code:`PMC` papers from a given date.
+- |Add| entrypoint :code:`bbs_database download`.
 - |Add| run the tox env ``check-apidoc`` in CI
 - |Add| tox environments ``apidoc`` and ``check-apidoc``
 - |Add| input type ``tei-xml`` for the ``bbs_database parse`` command.